Refactor main script and split modules
